requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
black==23.7.0
braceexpand==0.1.7
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
discord==2.3.2
discord-ext-bot==1.0.1
discord-py-slash-command==4.2.1
discord.py==2.3.2
ffmpeg==1.4
frozenlist==1.4.0
idna==3.4
import-expression==1.1.4
jishaku==2.5.1
line-profiler==4.0.3
multidict==6.0.4
mutagen==1.46.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pycparser==2.21
pycryptodomex==3.18.0
PyNaCl==1.5.0
python-dotenv==1.0.0
six==1.16.0
tomli==2.0.1
websockets==11.0.3
yarl==1.9.2
youtube-dl==2021.12.17
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz#sha256=8b5c0faca656bb8c2354d14dfd073aa188f64ec05224fdebd9a43ff78627266a
